Canvas View Implementation
Created a basic view with a fixed canvas size which will be the map size also for now.
A better rendering option can be implemented in the View later on.

All Entities are currently shown as a white square and when updated, they move +1 in both x and y directions.