Entity class with MovementComponent
Created an Entity that is displayable in the ViewCanvas. It also has a MovementSystem which is the basis for Steering behaviors and other stuff.

Added a new Button to spawn new Entities

Updated Vector tests as vector functions needed to be added for the class

Added Color Constants which for now have to be manually synced with the styles.css :root color scheme