Implemented Stats Component
Added a stats component which holds all kinds of stats like Health etc.

Created a general Stats list of stats I might use later on.

Seperated Entity from Food and created a new Class Monster for cleaner seperation because Monster has Health, but food doesnt. This is represented through a Alive Stat which basically is the stat that defines the condition for the isAlive() condition that decides if an ModelEntity is removed from the list inside the model.