RuleBuilder
Implemented Rules and a RuleBuilder which all can be added via a RuleManager. With Rules, advanced states can be represented like adding a decay Factor to the Health which didnt decay before but now that we are all hungry the health decays.

Basically made the decay values so small that Monsters have infitnite life, so they dont despawn too fast.