Visuals, Enhancements & More
Stats that depend on each other:
Stats can now depend on each other and currently the movement speed is depended on the size and the hunger gets higher faster when entities are smaller.

Fixes:
Fixed a graphical bug where the scaling would not apply perfectly. Now, the resolution is fixed and everything is displayed correctly.
Added Touch events to phone